commit 02:Final changes made for 02/01/2023
